requirements.txt
PyYAML
numpy
PySerial
PyVisa
pyvisa-py
uldaq
ipython
pyzmq
future
